refactor: 将下载逻辑提取到useDownload hook中
